Separate I/O methods from algorithm methods
