+++ include/hmcnc.h
@@ -74,7 +74,7 @@ struct Parameters {
 };
 
 // ----------------------
-// methods
+// Algorithm methods
 // ----------------------
 
 double BaumWelchEOnChrom(const std::vector<double> &startP,
@@ -153,6 +153,73 @@ double max_over_row(const std::vector<std::vector<double>> &v,
 double max_over_rows(const std::vector<std::vector<double>> &v, size_t col,
                      const std::vector<std::vector<double>> &v2, size_t nextState,
                      size_t nStates);
+
+// --------------------------
+// I/O methods
+// --------------------------
+
+void ReadCoverage(std::istream &covFile,
+                  const std::vector<std::string> &contigNames,
+                  std::vector<std::vector<int>> &covBins);
+void ReadCoverage(const std::string &covFileName,
+                  const std::vector<std::string> &contigNames,
+                  std::vector<std::vector<int>> &covBins);
+
+void ReadFai(std::istream &faiFile,
+             std::vector<std::string> &contigNames,
+             std::vector<int> &contigLengths);
+void ReadFai(const std::string faiFileName,
+             std::vector<std::string> &contigNames,
+             std::vector<int> &contigLengths);
+
+void ReadParameterFile(std::istream &file, int &nStates, double &covMean,
+                       double &covVar, int &maxState, int &maxCov,
+                       std::vector<double> &startP,
+                       std::vector<std::vector<double>> &transP,
+                       std::vector<std::vector<double>> &emisP);
+void ReadParameterFile(const std::string &fileName, int &nStates, double &covMean,
+                       double &covVar, int &maxState, int &maxCov,
+                       std::vector<double> &startP,
+                       std::vector<std::vector<double>> &transP,
+                       std::vector<std::vector<double>> &emisP);
+
+void ReadSNVs(std::istream &snvFile,
+              const std::vector<std::string> &contigNames,
+              std::vector<std::vector<SNV>> &snvs);
+void ReadSNVs(const std::string &snvFileName,
+              const std::vector<std::string> &contigNames,
+              std::vector<std::vector<SNV>> &snvs);
+
+void WriteCovBed(std::ostream &covFile,
+		             const std::vector<std::string> &contigNames,
+		             const std::vector<std::vector<int>> &covBins);
+void WriteCovBed(const std::string &covFileName,
+		             const std::vector<std::string> &contigNames,
+		             const std::vector<std::vector<int>> &covBins);
+
+void WriteParameterFile(std::ostream &file, int nStates, double covMean,
+                        double covVar, int maxState, int maxCov,
+                        const std::vector<double> &startP,
+                        const std::vector<std::vector<double>> &transP,
+                        const std::vector<std::vector<double>> &emisP);
+void WriteParameterFile(const std::string &fileName, int nStates, double covMean,
+                        double covVar, int maxState, int maxCov,
+                        const std::vector<double> &startP,
+                        const std::vector<std::vector<double>> &transP,
+                        const std::vector<std::vector<double>> &emisP);
+
+void WriteSNVs(std::ostream &snvFile,
+               const std::vector<std::string> &contigNames,
+               const std::vector<std::vector<SNV>> &snvs);
+void WriteSNVs(const std::string &snvFileName,
+               const std::vector<std::string> &contigNames,
+               const std::vector<std::vector<SNV>> &snvs);
+
+void WriteVCF(std::ostream &out, const std::string &refName,
+              const std::string &sampleName,
+              const std::vector<std::string> &contigNames,
+              const std::vector<int> &contigLengths,
+              const std::vector<std::vector<Interval>> &intervals);
 
 // --------------------------
 // main application runners
